fix: remove test file and fix pipeline
